+++ tests/test_cli.py
@@ -174,7 +174,7 @@ def test_can_invoke_execute_cli(
 )
 @pytest.mark.parametrize("amlignore", ("empty", "missing", "filled"))
 @pytest.mark.parametrize("gitignore", ("empty", "missing", "filled"))
-@pytest.mark.parametrize("extra_env", (([], {}), (["A=B", "C="], {'A': 'B', 'C': ''})))
+@pytest.mark.parametrize("extra_env", (([], {}), (["A=B", "C="], {"A": "B", "C": ""})))
 def test_can_invoke_run(
     patched_kedro_package,
     cli_context,
@@ -218,7 +218,7 @@ def test_can_invoke_run(
             ["-s", "subscription_id"]
             + (["--wait-for-completion"] if wait_for_completion else [])
             + (["--aml_env", aml_env] if aml_env else [])
-            + (sum([['--env-var', k] for k in extra_env[0]], [])),
+            + (sum([["--env-var", k] for k in extra_env[0]], [])),
             obj=cli_context,
         )
         assert result.exit_code == 0
@@ -238,8 +238,10 @@ def test_can_invoke_run(
             interactive_credentials.assert_not_called()
 
         created_pipeline = ml_client.jobs.create_or_update.call_args[0][0]
-        populated_env_vars = list(created_pipeline.jobs.values())[0].environment_variables
-        del populated_env_vars['KEDRO_AZURE_RUNNER_CONFIG']
+        populated_env_vars = list(created_pipeline.jobs.values())[
+            0
+        ].environment_variables
+        del populated_env_vars["KEDRO_AZURE_RUNNER_CONFIG"]
         assert populated_env_vars == extra_env[1]
 
 
@@ -329,6 +331,7 @@ def test_can_invoke_run_with_failed_pipeline(
         ml_client.compute.get.assert_called_once()
         ml_client.jobs.stream.assert_called_once()
 
+
 def test_fail_if_invalid_env_provided_in_run(
     patched_kedro_package,
     cli_context,
@@ -351,4 +354,7 @@ def test_fail_if_invalid_env_provided_in_run(
         runner = CliRunner()
         result = runner.invoke(cli.run, ["--env-var", "INVALID"], obj=cli_context)
         assert result.exit_code == 1
-        assert str(result.exception) == 'Invalid env-var: INVALID, expected format: KEY=VALUE'
+        assert (
+            str(result.exception)
+            == "Invalid env-var: INVALID, expected format: KEY=VALUE"
+        )